Create GTFS ingest errors debug endpoint, use Vec::from_iter()
GTFS debug will be used to create a leptos page detailing each broken feed and what's wrong with it.

As far as I can tell, there is little optimisation from Vec::from_iter() vs for in.
Source: https://users.rust-lang.org/t/performance-difference-between-iterator-and-for-loop/50254/4
It makes the code a bit cleaner, that's all.