Create struct for downloaded feed information
The structure `DownloadedFeedsInformation` in the file `src/next_db/transitland_import.rs` contains valuable information I intend to show to the user on a debug page using the `buttercup` repo.
If the operation is success and reingest is marked true, the parent function in `src/next_db/import.rs` shall insert into the database, and be ready to record attempted faults.

Now that I think about it, having a `failed` bool flag isn't a good idea because if the container vapourises, it doesn't know that it is okay.
A potential solution is to keep a zookeeper ephemeral node attached, and if it disconnects, mark the attempt as failed and clean up the other data tables.